+++ R/get_proj_denom_multi.R
@@ -10,7 +10,28 @@ get_proj_denom_multi <- function(Z, Y, trait_only = TRUE, trait_index = 1) {
         stop('`Z` and `Y` dimensions must match!  Got dim(Z) = ', toString(dim(Z)), ', dim(Y) = ', toString(dim(Y)) )
     
     # this is the generalized "denom"
+    # direct version, hoping this doesn't happen after fixed locus removal
     cov_mat <- solve( crossprod( Y, Z ) )
+    ## # "try" matrix inversion and try to recover if there is colinearity
+    ## cov_mat <- try( solve( crossprod( Y, Z ) ), silent = TRUE )
+    ## while ( "try-error" %in% class( cov_mat ) ) {
+    ##     ## # makes sense to throw away columns, one at the time
+    ##     ## print( Y )
+    ##     ## print( Z )
+    ##     # keep all rows
+    ##     # find index to remove
+    ##     k <- ncol( Y )
+    ##     if ( k == 1 ) {
+    ##         # NOTE: if we were down to a single column, that's always invertible (unless the dot product is zero), very unlikely (let's actually die if that happens)
+    ##         stop('Removed all columns trying to make `crossprod(Y, Z)` invertible!')
+    ##     } else {
+    ##         # remove columns now
+    ##         Y <- Y[ , -k ]
+    ##         Z <- Z[ , -k ]
+    ##         # try again!
+    ##         cov_mat <- try( solve( crossprod( Y, Z ) ) )
+    ##     }
+    ## }
 
     # this is the generalized projection matrix
     proj <- Z %*% cov_mat