+++ NEWS.md
@@ -14,3 +14,12 @@
 # 2020-07-14 - ligera 1.0.3.9000
 
 * Added multiscan wrapper for all versions of LIGERA (`ligera_multi`, `ligera2_multi`, `ligera2_bed_multi`), which iterates the genetic association scans and adds significant loci to covariates (options for one per iteration or all significant per iteration) to condition upon them and increase power on the weaker signals.
+
+# 2020-07-21 - ligera 1.0.4.9000
+
+* Added support for missing values in covariates matrix
+  * In each column of covariates matrix, missing values are replaced with the within-column mean value.
+  * All core versions updated (`ligera`, `ligera2`, and `ligera2_bed`).
+  * This was necessary for multiscan versions (`ligera_multi`, `ligera2_multi`, `ligera2_bed_multi`) to work when the genotype matrix has missing values, as these may be incorporated as covariates in the iteration.
+* Multiscan versions stop iterations if and when the number of covariates exceed the number of individuals, to avoid inverting singular matrices (problem becomes overdetermined).
+* Fixed `class` usage now that matrices return a two-element array in R version 4.